backend/requirements.txt
opencv-python~=4.11.0.86
dlib~=19.24.9
requests~=2.32.3
Flask~=3.1.0
Werkzeug~=3.1.3
itsdangerous~=2.2.0
numpy~=2.2.6
Authlib~=1.6.5
flask-cors~=6.0.0
PyJWT~=2.10.1
python-dotenv~=1.1.1